requirements.txt
aiohttp
aria2p==0.9.1
beautifulsoup4
bs4
cfscrape
cloudscraper
feedparser
hachoir
heroku3
html-telegraph-poster
hurry.filesize
imdbpy
js2py
lk21
lxml
messages
natsort
Pillow
psutil
psycopg2-binary
pycountry
pyrogram
python-dotenv
pytz
pyprog
python-telegram-bot
pybase64
requests
speedtest-cli
telegraph
tgcrypto
urllib3
yt-dlp
